本次修复了过滤规则，将basePageWidget的virtual FieldFilterProxyModel* createProxy(int status)=0;改为virtual FieldFilterProxyModel* createProxy(const QVariant& data)=0;支持后面的多种过滤

+++ ui/basepagewidget.cpp
@@ -168,7 +168,8 @@ void BasePageWidget::initTabs()
         m_tabBar->addTab(tabs[i].title);
 
         // 2. Proxy
-        auto proxy = createProxy(tabs[i].data.toInt());
+        auto proxy = createProxy(tabs[i].data);
+       // auto proxy = createProxy(tabs[i].data.toInt());
         if (!proxy || !m_model)
             continue;
